Add files via upload
After fitting the data to these models, a matlab data file (.mat) with all the inputs was generated. 

This package includes a total of 4 files; 4 for each of the 4 models. The names of these file are created using this system:
	Prefix = fitted1to3
	Sub-suffix = which model, i.e. full, agg, noncoop and stat
	Ending = x4 (denotes this is a global fit from 4 different chemical resonances). 
Where full = full model, agg = additive model, noncoop = non-cooperative and stat = statistical model.
As an example, the file name:
	fitted1to3noncoopx4.mat
corresponds to the data fitted to the non-cooperative binding model (equation S10).